fastapi-demo/requirements.txt
fastapi
uvicorn[standard]
aiogram
langchain
langchain-openai
langgraph
python-dotenv
playwright
pandas